services/armoriq-agent/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
pyyaml==6.0.1
armorclaw>=0.1.0
